refactor: improve the DX of model factories
