+++ resources/js/common/models/user.model.ts
@@ -1,3 +1,5 @@
+import { createFactory } from '@/test/createFactory';
+
 interface UserPreferences {
   prefersAbsoluteDates: boolean;
 }
@@ -10,3 +12,18 @@ export interface User {
   preferences: UserPreferences;
   unreadMessageCount: number;
 }
+
+export const createUser = createFactory<User>((faker) => {
+  const displayName = faker.internet.displayName();
+
+  return {
+    displayName,
+    avatarUrl: `http://media.retroachievements.org/UserPic/${displayName}.png`,
+    id: faker.number.int({ min: 1, max: 1000000 }),
+    legacyPermissions: faker.number.int({ min: 0, max: 4 }),
+    preferences: {
+      prefersAbsoluteDates: faker.datatype.boolean(),
+    },
+    unreadMessageCount: faker.number.int({ min: 0, max: 3 }),
+  };
+});
