+++ resources/js/features/forums/components/RecentPostsMainRoot/RecentPostsBreadcrumbs/index.ts
