+++ resources/js/features/forums/components/RecentPostsMainRoot/RecentPostsCards/index.ts
