+++ resources/js/features/forums/components/RecentPostsMainRoot/RecentPostsMainRoot.tsx
