+++ resources/js/features/forums/components/RecentPostsMainRoot/RecentPostsPagination/index.ts
