+++ resources/js/features/forums/components/RecentPostsMainRoot/RecentPostsTable/index.ts
