+++ resources/js/features/forums/models/index.ts
