--- resources/js/test/factories/index.ts
@@ -1,2 +1,0 @@
-export * from './recent-forum-post.factory';
-export * from './user.factory';
