+++ resources/js/test/index.ts
@@ -1,2 +1,1 @@
-export * from './factories';
 export * from './setup';
